Paul's edits
* Don't ask for changes for all conversion specs; only for %z and %Z.
The other conversion specifications don't have these issues.

* Also ask for clarity as to whether %z and %Z's output can depend on
global state. The argument that output can depend only on tm_isdst also
argues against depending on global state.

* Make it clearer which relevant struct tm / strftime features are in
older POSIX, and which are in POSIX.1-2024, as the C committee members
may not be familiar with the details.

* Give examples of real-world platforms affected by this issue.

* Also give some example C code illustrating the issue without resorting
to `#ifdef POSIX`. This will help show why this is a C issue, not merely
a POSIX issue.

* Also give an example using a timezone people are more used to, like
America/New_York or Europe/London. This may help convince people that
the problems are not as obscure as one might think.

* No need to include <stdlib.h> in the Caracas example.

* The Caracas example should set tm_isdst to 0, not 1, because there is
no daylight saving time there.

* It's better to discuss the issue of %z and %Z having undefined
separately from the Caracas example. That is, the Caracas example should
talk only about ambiguous timestamps and not try to do too much; the
undefined behavior stuff can be discussed separately.

* C23 style seems to be to say `strftime` not `strftime()` when talking
about the function.

* Assuming we ask for changes only for %z and %Z, Proposal 1 can become
a lot more modest. It doesn't need to change the general wording, only
the wording for %z and for %Z.